adds deletion of old images when deleting or changing a product

+++ MongoDB/controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require("../util/file");
+
 const { validationResult } = require("express-validator/check");
 
 const Product = require("../models/product");
@@ -131,7 +133,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      if (image) product.imageUrl = image.path;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
 
       return product.save().then((result) => {
         console.log("Updated Product.");
@@ -165,10 +170,17 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    // Product.findByIdAndRemove(prodId)
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) return next(new Error("Product not found."));
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then((result) => {
       console.log("Removed Product.");
       res.redirect("/admin/products");
-    });
+    })
+    .catch((err) => next(err));
+
+  // Product.findByIdAndRemove(prodId)
 };
